STT/realtime_server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Note: For AssemblyAI's official SDK (optional alternative to our WebSocket implementation)
# assemblyai>=0.25.0
